Add 4.12 -> 4.13 upgrade testing for WMCO (#43490)
This copies the workflow that is used for the 4.13->4.14 upgrade, using
Azure instead of vSphere. Azure is being used as we wish to test storage
migration which occurs only for Azure across this upgrade.

The operator-sdk version was bumped to include this required fix:
https://github.com/operator-framework/operator-sdk/pull/6477

`make update` was ran to generate the presubmit file