OPNET-359: Remove unnecessary second operator job (#43725)
There's nothing SDN-specific about the operator job so there's no
need to run it twice on each PR. This removes the SDN version of
the job since we primarily care about OVNK at this point anyway.

Note that we already don't run the operator job in periodics, so
no changes were necessary there.