add keda e2e tests and component image builds (#43390)
This adds:
- image builds for the keda components (controller, metrics adapter,
  webhooks)
- an image build deriving from the upstream keda-tools image that the
test suite expects to run in
- a test that runs the keda e2e suite against an aws cluster

The keda test suite was not designed to run in OpenShift CI. It mostly
expects to be run on a "vanilla kube" cluster, so there are some
quirks, mostly around privilege/permissions, the helm charts that it
tries to install, and the myriad public images it tries to pull.

The goal here is just to get some e2e tests running so we have some
signal and then clean this up over time, potentially with some upstream
test refactoring.

The scalers/cpu and internal/nternals/global_custom_ca_test are
currently not working and have been intentionally omitted, but they will
be re-added later once we figure out why.