hack: allow building on systems with custom container config (#44507)
Currently the build scripts allow configuring custom container engine
but do not allow configuring custom config parameters that would be
passed to every invocation of `CONTAINER_ENGINE pull` or
`CONTAINER_ENGINE run`. As a result it is not possible to use e.g.
system with a container engine storing authfile in a non-default
location.

Even though Makefile allows specifying `CONTAINER_ENGINE_OPTS`, this
value is not used in all the scripts but only in some of them.

With this PR the value of `CONTAINER_ENGINE_OPTS` will be carried
everywhere where we pull or run a container.

Sample use that this PR enables

```
CONTAINER_ENGINE=podman
CONTAINER_ENGINE_OPTS="--authfile /tmp/auth.json --platform linux/amd64"
make update
```