ARMOCP-567: AWS and Azure test configs for clusters heterogeneous in arch and kernel pagesize (#45768)
* Fix step mco-conf-day2-create-custom-mcp

- Do not double quote the label selector as oc would interpret it as unique
- Run set_proxy() before validate_params()
- Minor harmless improvements in the code style

* ARMOCP-567: AWS and Azure test configs for clusters heterogeneous in arch and kernel pagesize

This commit adds an AWS connected and Azure disconnected test configs to
run openshift-tests-private against clusters heterogeneous in CPU
Architecture and kernel pagesize.

The env consists of:
- 3 arm64 masters
- 2 arm64 workers with 4k pagesize kernel
- 1 arm64 worker with 64k pagesize kernel
- 1 amd64 worker (only 4k pagesize kernel is supported)

Frequency is set to f28 for both the test configs. Upgrade jobs will be
added once the config file from 4.15 stable to 4.15 nightly is available.